[#186] fixed distribution across files and added base class for reusability
